+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -98,6 +98,14 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`bugprone-incorrect-enable-shared-from-this
+  <clang-tidy/checks/bugprone/incorrect-enable-shared-from-this>` check.
+
+  Check if class/structs publicly inherits from ``std::enable_shared_from_this``,
+  because otherwise when ``shared_from_this`` is called, the code will throw 
+  ``std::bad_weak_ptr``. Also checks for ``std::`` in 
+  ``std::enable_shared_from_this``.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^